+++ src/index.js
@@ -1,0 +1,6 @@
+import State from './state';
+import { syncEntityEvents } from './state.functions';
+
+State.syncEntityEvents = syncEntityEvents;
+
+export default State;