+++ test/spec/state.spec.js
@@ -3,33 +3,9 @@ describe('State', function () {
     this.state = new Mn.State();
   });
 
-  describe('when instantiating with the component option', function () {
+  describe('when', function () {
 
-    it('should self-destruct when component fires destroy event', function () {
-    });
-
-    it('should bind componentEvents to component', function () {
-    });
-  });
-
-  describe('when instantiating with the initialState option', function () {
-
-    it('the state model attributes should match initialState', function () {
-    });
-  });
-
-  describe('when calling setComponent with component argument', function () {
-
-    it('should self-destruct when component fires destroy event', function () {
-    });
-
-    it('should bind componentEvents to component', function () {
-    });
-  });
-
-  describe('when calling setState with attrs argument', function () {
-
-    it('the state model attributes should match attrs', function () {
+    it('should', function () {
     });
   });
 });